requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.9.1
attrs==25.3.0
boto3==1.40.33
botocore==1.40.33
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
dj-database-url==3.0.1
Django==5.2.6
django-cors-headers==4.8.0
django-environ==0.12.0
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
mailjet-rest==1.5.1
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
pycparser==2.23
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
s3transfer==0.14.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
